Moved response and request classes to seperate folders for organization
