+++ app/src/main/java/com/sethpthomas91/httpserver/ServerLogic.java
@@ -2,6 +2,10 @@ package com.sethpthomas91.httpserver;
 
 import com.sethpthomas91.httpserver.interfaces.HttpRequestInterface;
 import com.sethpthomas91.httpserver.interfaces.ServerLogicInterface;
+import com.sethpthomas91.httpserver.response.Body;
+import com.sethpthomas91.httpserver.response.Header;
+import com.sethpthomas91.httpserver.response.HttpResponseWrapper;
+import com.sethpthomas91.httpserver.response.StatusLine;
 
 import java.util.ArrayList;
 import java.util.List;
