+++ app/src/main/java/com/sethpthomas91/httpserver/interfaces/HttpRequestInterface.java
@@ -1,6 +1,6 @@
 package com.sethpthomas91.httpserver.interfaces;
 
-import com.sethpthomas91.httpserver.RequestLine;
+import com.sethpthomas91.httpserver.request.RequestLine;
 
 public interface HttpRequestInterface {
     RequestLine getRequestLine();
