+++ app/src/main/java/com/sethpthomas91/httpserver/response/Header.java
@@ -1,4 +1,4 @@
-package com.sethpthomas91.httpserver;
+package com.sethpthomas91.httpserver.response;
 
 public class Header {
     String CRLF = "\r\n";
