+++ app/src/test/java/com/sethpthomas91/httpserver/mocks/MockServerLogic.java
@@ -1,7 +1,6 @@
 package com.sethpthomas91.httpserver.mocks;
 
-import com.sethpthomas91.httpserver.HttpRequestWrapper;
-import com.sethpthomas91.httpserver.HttpResponseWrapper;
+import com.sethpthomas91.httpserver.response.HttpResponseWrapper;
 import com.sethpthomas91.httpserver.interfaces.HttpRequestInterface;
 import com.sethpthomas91.httpserver.interfaces.ServerLogicInterface;
